+++ ext_localconf.php
@@ -40,19 +40,14 @@ else {
 }
 
 if (TYPO3_MODE == 'BE') {
-	// Einbindung einer PageTSConfig
-	tx_rnbase_util_Extensions::addPageTSConfig('<INCLUDE_TYPOSCRIPT: source="FILE:EXT:cfc_league/mod1/pageTSconfig.txt">');
+    // Einbindung einer PageTSConfig
+    tx_rnbase_util_Extensions::addPageTSConfig('<INCLUDE_TYPOSCRIPT: source="FILE:EXT:cfc_league/mod1/pageTSconfig.txt">');
 
-	// Not working with
-// 	// Register form element for logofield in team record
-// 	if(tx_rnbase_util_TYPO3::isTYPO76OrHigher()) {
-// 		tx_rnbase::load('Tx_Cfcleague_Form_SelectTeamLogo');
-// 		$GLOBALS['TYPO3_CONF_VARS']['SYS']['formEngine']['nodeRegistry'][7493905268] = array(
-// 				'nodeName' => 't3s_teamlogo',
-// 				'priority' => 40,
-// 				'class' => \Tx_Cfcleague_Form_SelectTeamLogo::class,
-// 		);
-// 	}
+    $GLOBALS['TYPO3_CONF_VARS']['SYS']['formEngine']['nodeRegistry']['t3sports_logoSelect'] = [
+        'nodeName' => 't3sLogoSelect',
+        'priority' => '70',
+        'class' => 'System25\T3sports\Form\Element\LogoSelect',
+    ];
 }